xray-classification/requirements.txt
flask==1.1.1
msgpack==0.6.2
numpy==1.18.0
opencv-python==4.1.2.30
pillow==7.0.0
pydicom==1.3.0
waitress==1.4.2
pytorch==1.4.0
skimage==0.14.1
easydict